[refactor] modify getUserMatchInfo function name, change getAge function and ES setting to other folder
